WIP: progress toward making groupby work with multiple arguments
It definitely doesn't work properly yet, totally mixing up coordinates,
data variables and multi-indexes.
