+++ xarray/test/test_dataarray.py
@@ -1131,270 +1131,6 @@ class TestDataArray(TestCase):
         actual = a.groupby('b').fillna(DataArray([0, 2], dims='b'))
         self.assertDataArrayIdentical(expected, actual)
 
-    def test_groupby_iter(self):
-        for ((act_x, act_dv), (exp_x, exp_ds)) in \
-                zip(self.dv.groupby('y'), self.ds.groupby('y')):
-            self.assertEqual(exp_x, act_x)
-            self.assertDataArrayIdentical(exp_ds['foo'], act_dv)
-        for ((_, exp_dv), act_dv) in zip(self.dv.groupby('x'), self.dv):
-            self.assertDataArrayIdentical(exp_dv, act_dv)
-
-    def make_groupby_example_array(self):
-        da = self.dv.copy()
-        da.coords['abc'] = ('y', np.array(['a'] * 9 + ['c'] + ['b'] * 10))
-        da.coords['y'] = 20 + 100 * da['y']
-        return da
-
-    def test_groupby_properties(self):
-        grouped = self.make_groupby_example_array().groupby('abc')
-        expected_unique = Variable('abc', ['a', 'b', 'c'])
-        self.assertVariableEqual(expected_unique, grouped.unique_coord)
-        self.assertEqual(3, len(grouped))
-
-    def test_groupby_apply_identity(self):
-        expected = self.make_groupby_example_array()
-        idx = expected.coords['y']
-
-        def identity(x):
-            return x
-
-        for g in ['x', 'y', 'abc', idx]:
-            for shortcut in [False, True]:
-                for squeeze in [False, True]:
-                    grouped = expected.groupby(g, squeeze=squeeze)
-                    actual = grouped.apply(identity, shortcut=shortcut)
-                    self.assertDataArrayIdentical(expected, actual)
-
-    def test_groupby_sum(self):
-        array = self.make_groupby_example_array()
-        grouped = array.groupby('abc')
-
-        expected_sum_all = Dataset(
-            {'foo': Variable(['abc'], np.array([self.x[:, :9].sum(),
-                                                self.x[:, 10:].sum(),
-                                                self.x[:, 9:10].sum()]).T),
-             'abc': Variable(['abc'], np.array(['a', 'b', 'c']))})['foo']
-        self.assertDataArrayAllClose(expected_sum_all, grouped.reduce(np.sum))
-        self.assertDataArrayAllClose(expected_sum_all, grouped.sum())
-
-        expected = DataArray([array['y'].values[idx].sum() for idx
-                              in [slice(9), slice(10, None), slice(9, 10)]],
-                             [['a', 'b', 'c']], ['abc'])
-        actual = array['y'].groupby('abc').apply(np.sum)
-        self.assertDataArrayAllClose(expected, actual)
-        actual = array['y'].groupby('abc').sum()
-        self.assertDataArrayAllClose(expected, actual)
-
-        expected_sum_axis1 = Dataset(
-            {'foo': (['x', 'abc'], np.array([self.x[:, :9].sum(1),
-                                             self.x[:, 10:].sum(1),
-                                             self.x[:, 9:10].sum(1)]).T),
-             'x': self.ds['x'],
-             'abc': Variable(['abc'], np.array(['a', 'b', 'c']))})['foo']
-        self.assertDataArrayAllClose(expected_sum_axis1,
-                                     grouped.reduce(np.sum, 'y'))
-        self.assertDataArrayAllClose(expected_sum_axis1, grouped.sum('y'))
-
-    def test_groupby_count(self):
-        array = DataArray([0, 0, np.nan, np.nan, 0, 0],
-                          coords={'cat': ('x', ['a', 'b', 'b', 'c', 'c', 'c'])},
-                          dims='x')
-        actual = array.groupby('cat').count()
-        expected = DataArray([1, 1, 2], coords=[('cat', ['a', 'b', 'c'])])
-        self.assertDataArrayIdentical(actual, expected)
-
-    @unittest.skip('needs to be fixed for shortcut=False, keep_attrs=False')
-    def test_groupby_reduce_attrs(self):
-        array = self.make_groupby_example_array()
-        array.attrs['foo'] = 'bar'
-
-        for shortcut in [True, False]:
-            for keep_attrs in [True, False]:
-                print('shortcut=%s, keep_attrs=%s' % (shortcut, keep_attrs))
-                actual = array.groupby('abc').reduce(
-                    np.mean, keep_attrs=keep_attrs, shortcut=shortcut)
-                expected = array.groupby('abc').mean()
-                if keep_attrs:
-                    expected.attrs['foo'] = 'bar'
-                self.assertDataArrayIdentical(expected, actual)
-
-    def test_groupby_apply_center(self):
-        def center(x):
-            return x - np.mean(x)
-
-        array = self.make_groupby_example_array()
-        grouped = array.groupby('abc')
-
-        expected_ds = array.to_dataset()
-        exp_data = np.hstack([center(self.x[:, :9]),
-                              center(self.x[:, 9:10]),
-                              center(self.x[:, 10:])])
-        expected_ds['foo'] = (['x', 'y'], exp_data)
-        expected_centered = expected_ds['foo']
-        self.assertDataArrayAllClose(expected_centered, grouped.apply(center))
-
-    def test_groupby_apply_ndarray(self):
-        # regression test for #326
-        array = self.make_groupby_example_array()
-        grouped = array.groupby('abc')
-        actual = grouped.apply(np.asarray)
-        self.assertDataArrayEqual(array, actual)
-
-    def test_groupby_apply_changes_metadata(self):
-        def change_metadata(x):
-            x.coords['x'] = x.coords['x'] * 2
-            x.attrs['fruit'] = 'lemon'
-            return x
-
-        array = self.make_groupby_example_array()
-        grouped = array.groupby('abc')
-        actual = grouped.apply(change_metadata)
-        expected = array.copy()
-        expected = change_metadata(expected)
-        self.assertDataArrayEqual(expected, actual)
-
-    def test_groupby_math(self):
-        array = self.make_groupby_example_array()
-        for squeeze in [True, False]:
-            grouped = array.groupby('x', squeeze=squeeze)
-
-            expected = array + array.coords['x']
-            actual = grouped + array.coords['x']
-            self.assertDataArrayIdentical(expected, actual)
-
-            actual = array.coords['x'] + grouped
-            self.assertDataArrayIdentical(expected, actual)
-
-            ds = array.coords['x'].to_dataset('X')
-            expected = array + ds
-            actual = grouped + ds
-            self.assertDatasetIdentical(expected, actual)
-
-            actual = ds + grouped
-            self.assertDatasetIdentical(expected, actual)
-
-        grouped = array.groupby('abc')
-        expected_agg = (grouped.mean() - np.arange(3)).rename(None)
-        actual = grouped - DataArray(range(3), [('abc', ['a', 'b', 'c'])])
-        actual_agg = actual.groupby('abc').mean()
-        self.assertDataArrayAllClose(expected_agg, actual_agg)
-
-        with self.assertRaisesRegexp(TypeError, 'only support binary ops'):
-            grouped + 1
-        with self.assertRaisesRegexp(TypeError, 'only support binary ops'):
-            grouped + grouped
-        with self.assertRaisesRegexp(TypeError, 'in-place operations'):
-            array += grouped
-
-    def test_groupby_math_not_aligned(self):
-        array = DataArray(range(4), {'b': ('x', [0, 0, 1, 1])}, dims='x')
-        other = DataArray([10], dims='b')
-        actual = array.groupby('b') + other
-        expected = DataArray([10, 11, np.nan, np.nan], array.coords)
-        self.assertDataArrayIdentical(expected, actual)
-
-        other = DataArray([10], coords={'c': 123}, dims='b')
-        actual = array.groupby('b') + other
-        expected.coords['c'] = (['x'], [123] * 2 + [np.nan] * 2)
-        self.assertDataArrayIdentical(expected, actual)
-
-        other = Dataset({'a': ('b', [10])})
-        actual = array.groupby('b') + other
-        expected = Dataset({'a': ('x', [10, 11, np.nan, np.nan])},
-                           array.coords)
-        self.assertDatasetIdentical(expected, actual)
-
-    def test_groupby_restore_dim_order(self):
-        array = DataArray(np.random.randn(5, 3),
-                          coords={'a': ('x', range(5)), 'b': ('y', range(3))},
-                          dims=['x', 'y'])
-        for by, expected_dims in [('x', ('x', 'y')),
-                                  ('y', ('x', 'y')),
-                                  ('a', ('a', 'y')),
-                                  ('b', ('x', 'b'))]:
-            result = array.groupby(by).apply(lambda x: x.squeeze())
-            self.assertEqual(result.dims, expected_dims)
-
-    def test_groupby_first_and_last(self):
-        array = DataArray([1, 2, 3, 4, 5], dims='x')
-        by = DataArray(['a'] * 2 + ['b'] * 3, dims='x', name='ab')
-
-        expected = DataArray([1, 3], [('ab', ['a', 'b'])])
-        actual = array.groupby(by).first()
-        self.assertDataArrayIdentical(expected, actual)
-
-        expected = DataArray([2, 5], [('ab', ['a', 'b'])])
-        actual = array.groupby(by).last()
-        self.assertDataArrayIdentical(expected, actual)
-
-        array = DataArray(np.random.randn(5, 3), dims=['x', 'y'])
-        expected = DataArray(array[[0, 2]], {'ab': ['a', 'b']}, ['ab', 'y'])
-        actual = array.groupby(by).first()
-        self.assertDataArrayIdentical(expected, actual)
-
-        actual = array.groupby('x').first()
-        expected = array  # should be a no-op
-        self.assertDataArrayIdentical(expected, actual)
-
-    def make_groupby_multidim_example_array(self):
-        return DataArray([[[0,1],[2,3]],[[5,10],[15,20]]],
-                        coords={'lon': (['ny', 'nx'], [[30., 40.], [40., 50.]] ),
-                                'lat': (['ny', 'nx'], [[10., 10.], [20., 20.]] ),},
-                        dims=['time', 'ny', 'nx'])
-
-    def test_groupby_multidim(self):
-        array = self.make_groupby_multidim_example_array()
-        for dim, expected_sum in [
-                ('lon', DataArray([5, 28, 23], coords={'lon': [30., 40., 50.]})),
-                ('lat', DataArray([16, 40], coords={'lat': [10., 20.]}))]:
-            actual_sum = array.groupby(dim).sum()
-            self.assertDataArrayIdentical(expected_sum, actual_sum)
-
-    def test_groupby_multidim_apply(self):
-        array = self.make_groupby_multidim_example_array()
-        actual = array.groupby('lon').apply(
-                lambda x : x - x.mean(), shortcut=False)
-        expected = DataArray([[[-2.5, -6.], [-5., -8.5]],
-                              [[ 2.5,  3.], [ 8.,  8.5]]],
-                    coords=array.coords, dims=array.dims)
-        self.assertDataArrayIdentical(expected, actual)
-
-    def test_groupby_bins(self):
-        array = DataArray(np.arange(4), dims='dim_0')
-        # the first value should not be part of any group ("right" binning)
-        array[0] = 99
-        # bins follow conventions for pandas.cut
-        # http://pandas.pydata.org/pandas-docs/stable/generated/pandas.cut.html
-        bins = [0,1.5,5]
-        bin_coords = ['(0, 1.5]', '(1.5, 5]']
-        expected = DataArray([1,5], dims='dim_0_bins',
-                        coords={'dim_0_bins': bin_coords})
-        # the problem with this is that it overwrites the dimensions of array!
-        #actual = array.groupby('dim_0', bins=bins).sum()
-        actual = array.groupby_bins('dim_0', bins).apply(
-                                    lambda x : x.sum(), shortcut=False)
-        self.assertDataArrayIdentical(expected, actual)
-        # make sure original array dims are unchanged
-        # (would fail with shortcut=True above)
-        self.assertEqual(len(array.dim_0), 4)
-
-    def test_groupby_bins_multidim(self):
-        array = self.make_groupby_multidim_example_array()
-        bins = [0,15,20]
-        bin_coords = ['(0, 15]', '(15, 20]']
-        expected = DataArray([16, 40], dims='lat_bins',
-                                coords={'lat_bins': bin_coords})
-        actual = array.groupby_bins('lat', bins).apply(
-                                    lambda x : x.sum(), shortcut=False)
-        self.assertDataArrayIdentical(expected, actual)
-        # modify the array coordinates to be non-monotonic after unstacking
-        array['lat'].data = np.array([[10., 20.], [20., 10.]])
-        expected = DataArray([28, 28], dims='lat_bins',
-                    coords={'lat_bins': bin_coords})
-        actual = array.groupby_bins('lat', bins).apply(
-                                    lambda x : x.sum(), shortcut=False)
-        self.assertDataArrayIdentical(expected, actual)
-
     def make_rolling_example_array(self):
         times = pd.date_range('2000-01-01', freq='1D', periods=21)
         values = np.random.random((21, 4))
