+++ xarray/test/test_dataset.py
@@ -1499,141 +1499,6 @@ class TestDataset(TestCase):
         with self.assertRaisesRegexp(ValueError, 'cannot select a dimension'):
             data.squeeze('y')
 
-    def test_groupby(self):
-        data = Dataset({'z': (['x', 'y'], np.random.randn(3, 5))},
-                       {'x': ('x', list('abc')),
-                        'c': ('x', [0, 1, 0])})
-        groupby = data.groupby('x')
-        self.assertEqual(len(groupby), 3)
-        expected_groups = {'a': 0, 'b': 1, 'c': 2}
-        self.assertEqual(groupby.groups, expected_groups)
-        expected_items = [('a', data.isel(x=0)),
-                          ('b', data.isel(x=1)),
-                          ('c', data.isel(x=2))]
-        for actual, expected in zip(groupby, expected_items):
-            self.assertEqual(actual[0], expected[0])
-            self.assertDatasetEqual(actual[1], expected[1])
-
-        identity = lambda x: x
-        for k in ['x', 'c', 'y']:
-            actual = data.groupby(k, squeeze=False).apply(identity)
-            self.assertDatasetEqual(data, actual)
-
-    def test_groupby_returns_new_type(self):
-        data = Dataset({'z': (['x', 'y'], np.random.randn(3, 5))})
-
-        actual = data.groupby('x').apply(lambda ds: ds['z'])
-        expected = data['z']
-        self.assertDataArrayIdentical(expected, actual)
-
-        actual = data['z'].groupby('x').apply(lambda x: x.to_dataset())
-        expected = data
-        self.assertDatasetIdentical(expected, actual)
-
-    def test_groupby_iter(self):
-        data = create_test_data()
-        for n, (t, sub) in enumerate(list(data.groupby('dim1'))[:3]):
-            self.assertEqual(data['dim1'][n], t)
-            self.assertVariableEqual(data['var1'][n], sub['var1'])
-            self.assertVariableEqual(data['var2'][n], sub['var2'])
-            self.assertVariableEqual(data['var3'][:, n], sub['var3'])
-
-    def test_groupby_errors(self):
-        data = create_test_data()
-        with self.assertRaisesRegexp(ValueError, 'must have a name'):
-            data.groupby(np.arange(10))
-        with self.assertRaisesRegexp(ValueError, 'length does not match'):
-            data.groupby(data['dim1'][:3])
-        with self.assertRaisesRegexp(ValueError, "must have a 'dims'"):
-            data.groupby(data.coords['dim1'].to_index())
-
-    def test_groupby_reduce(self):
-        data = Dataset({'xy': (['x', 'y'], np.random.randn(3, 4)),
-                        'xonly': ('x', np.random.randn(3)),
-                        'yonly': ('y', np.random.randn(4)),
-                        'letters': ('y', ['a', 'a', 'b', 'b'])})
-
-        expected = data.mean('y')
-        expected['yonly'] = expected['yonly'].variable.expand_dims({'x': 3})
-        actual = data.groupby('x').mean()
-        self.assertDatasetAllClose(expected, actual)
-
-        actual = data.groupby('x').mean('y')
-        self.assertDatasetAllClose(expected, actual)
-
-        letters = data['letters']
-        expected = Dataset({'xy': data['xy'].groupby(letters).mean(),
-                            'xonly': (data['xonly'].mean().variable
-                                      .expand_dims({'letters': 2})),
-                            'yonly': data['yonly'].groupby(letters).mean()})
-        actual = data.groupby('letters').mean()
-        self.assertDatasetAllClose(expected, actual)
-
-    def test_groupby_math(self):
-        reorder_dims = lambda x: x.transpose('dim1', 'dim2', 'dim3', 'time')
-
-        ds = create_test_data()
-        for squeeze in [True, False]:
-            grouped = ds.groupby('dim1', squeeze=squeeze)
-
-            expected = reorder_dims(ds + ds.coords['dim1'])
-            actual = grouped + ds.coords['dim1']
-            self.assertDatasetIdentical(expected, reorder_dims(actual))
-
-            actual = ds.coords['dim1'] + grouped
-            self.assertDatasetIdentical(expected, reorder_dims(actual))
-
-            ds2 = 2 * ds
-            expected = reorder_dims(ds + ds2)
-            actual = grouped + ds2
-            self.assertDatasetIdentical(expected, reorder_dims(actual))
-
-            actual = ds2 + grouped
-            self.assertDatasetIdentical(expected, reorder_dims(actual))
-
-        grouped = ds.groupby('numbers')
-        zeros = DataArray([0, 0, 0, 0], [('numbers', range(4))])
-        expected = ((ds + Variable('dim3', np.zeros(10)))
-                    .transpose('dim3', 'dim1', 'dim2', 'time'))
-        actual = grouped + zeros
-        self.assertDatasetEqual(expected, actual)
-
-        actual = zeros + grouped
-        self.assertDatasetEqual(expected, actual)
-
-        with self.assertRaisesRegexp(ValueError, 'dimensions .* do not exist'):
-            grouped + ds
-        with self.assertRaisesRegexp(ValueError, 'dimensions .* do not exist'):
-            ds + grouped
-        with self.assertRaisesRegexp(TypeError, 'only support binary ops'):
-            grouped + 1
-        with self.assertRaisesRegexp(TypeError, 'only support binary ops'):
-            grouped + grouped
-        with self.assertRaisesRegexp(TypeError, 'in-place operations'):
-            ds += grouped
-
-        ds = Dataset({'x': ('time', np.arange(100)),
-                      'time': pd.date_range('2000-01-01', periods=100)})
-        with self.assertRaisesRegexp(ValueError, 'incompat.* grouped binary'):
-            ds + ds.groupby('time.month')
-
-    def test_groupby_math_virtual(self):
-        ds = Dataset({'x': ('t', [1, 2, 3])},
-                     {'t': pd.date_range('20100101', periods=3)})
-        grouped = ds.groupby('t.day')
-        actual = grouped - grouped.mean()
-        expected = Dataset({'x': ('t', [0, 0, 0])},
-                           ds[['t', 't.day']])
-        self.assertDatasetIdentical(actual, expected)
-
-    def test_groupby_nan(self):
-        # nan should be excluded from groupby
-        ds = Dataset({'foo': ('x', [1, 2, 3, 4])},
-                     {'bar': ('x', [1, 1, 2, np.nan])})
-        actual = ds.groupby('bar').mean()
-        expected = Dataset({'foo': ('bar', [1.5, 3]), 'bar': [1, 2]})
-        self.assertDatasetIdentical(actual, expected)
-
     def test_resample_and_first(self):
         times = pd.date_range('2000-01-01', freq='6H', periods=10)
         ds = Dataset({'foo': (['time', 'x', 'y'], np.random.randn(10, 5, 3)),
